落库口径登记表规划的两项改动：请假支持半天、工时统一按半小时取整
- ComputeWorkHours 出口从 Math.Round(x,2) 改成 FloorToHalf(x)，全链路（本地打卡/设备同步/
  补卡重算/月度汇总自愈补算/请假时长）统一到半小时精度；模板月度汇总表日格子同步改整数为
  半小时，跟自己的合计列口径一致；ExcelExportHelper 删掉重复的 HalfFloor 实现
- 新增 ApplyLeaveHoursCap/ResolveDailyStandardHours/ResolveLeaveDaysFraction 三个公共静态
  方法，让"请假当天如果还有打卡"按"标准工时-已批准请假小时数"封顶结算工时，不再是整天
  无条件清零；五处写入点（本地打卡、设备同步、补卡重算、月度汇总自愈补算、审批回写）统一改造
- LeaveHours 回写从覆盖(=)改成累加(+=)，避免同一天多张假单互相顶掉
- ZKDeviceSyncService 补充注入 IOptions<AppSettingsOptions>，取默认标准工时时不再和其它三条
  路径口径分裂
- 新增/补充单元测试锁定上述行为（LeaveHoursTests、MidCheckWindowLogicTests）

注意：LeaveDays 折算规则里"占比>=0.5算1天"这条边界，会让最常见的"请假4小时/标准工时8小时"
场景算成1整天而不是0.5天，已按文档字面实现并测试锁定，但怀疑不是想要的效果，需要确认后再定。

+++ Models/DTOs/AttendanceDTOs.cs
@@ -159,9 +159,10 @@ public class TemplateReportRowDto
     public int     NightShiftDays  { get; set; }   // 夜班天数
     public decimal NightShiftHours { get; set; }   // 夜班总工时（小时，口径同 TotalWorkHours：按半小时取整累加）
 
-    /// <summary>每天的工时（整数，舍去小数）；当天没有工时（休息/请假/旷工等）为 null，导出时显示空白。
+    /// <summary>每天的工时（按半小时取整，跟月度合计同一个取整口径，不再是"整数舍去小数"）；
+    /// 当天没有工时（休息/请假/旷工等）为 null，导出时显示空白。
     /// 下标和 TemplateReportResultDto.Dates 一一对应。</summary>
-    public List<int?> DailyHours { get; set; } = [];
+    public List<decimal?> DailyHours { get; set; } = [];
 
     /// <summary>每天是不是上的夜班（跨天班次，或没排班时按打卡时间兜底判断）；导出 Excel 时用来把当天格子标黄。
     /// 下标和 DailyHours/TemplateReportResultDto.Dates 一一对应。</summary>
